moving layers to each seperate file
